Fix installation error due to SHOW COLUMNS
This patch was provided by @r4cdv as a fix for the error described
in #7.  I (Kevin) have been unable to reproduce the error and I do not
understand how the codepath introduced by this patch differs from the
[existing codepath for SHOW COLUMNS][show-columns] [entered below the
proposed patch][enter-installing].

For that reason, this commit will not be merged into the current
development branch.  It is being kept on a separate branch so that if
other users experience the same problem they can help evaluate this fix.

[show-columns]: https://github.com/kevinoid/postgresql-for-wordpress/blob/68d8ead5adec822d37d95aed012eefb3447ea04b/pg4wp/driver_pgsql_install.php#L55
[enter-installing]: https://github.com/kevinoid/postgresql-for-wordpress/blob/68d8ead5adec822d37d95aed012eefb3447ea04b/pg4wp/driver_pgsql.php#L496

Fixes #7